moved insuranceInformation and InsuranceAttestation out of common folder
